refactor: 도메인 구조 통일
