+++ api/src/main/java/com/tdns/toks/api/domain/user/service/UserApiService.java
@@ -2,10 +2,10 @@ package com.tdns.toks.api.domain.user.service;
 
 import com.tdns.toks.core.common.type.JwtToken;
 import com.tdns.toks.core.config.security.JwtTokenProvider;
-import com.tdns.toks.core.domain.model.entity.User;
-import com.tdns.toks.core.domain.service.UserService;
-import com.tdns.toks.core.domain.type.UserRole;
-import com.tdns.toks.core.domain.type.UserStatus;
+import com.tdns.toks.core.domain.user.model.entity.User;
+import com.tdns.toks.core.domain.user.service.UserService;
+import com.tdns.toks.core.domain.user.type.UserRole;
+import com.tdns.toks.core.domain.user.type.UserStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
