+++ core/src/main/java/com/tdns/toks/core/common/filter/JwtAuthenticationFilter.java
@@ -3,9 +3,9 @@ package com.tdns.toks.core.common.filter;
 import com.tdns.toks.core.common.exception.ApplicationErrorType;
 import com.tdns.toks.core.common.exception.SilentApplicationErrorException;
 import com.tdns.toks.core.config.security.JwtTokenProvider;
-import com.tdns.toks.core.domain.model.dto.UserDTO;
-import com.tdns.toks.core.domain.model.entity.User;
-import com.tdns.toks.core.domain.service.UserService;
+import com.tdns.toks.core.domain.user.model.dto.UserDTO;
+import com.tdns.toks.core.domain.user.model.entity.User;
+import com.tdns.toks.core.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
