+++ core/src/main/java/com/tdns/toks/core/common/service/UserDetailService.java
@@ -2,9 +2,9 @@ package com.tdns.toks.core.common.service;
 
 import com.tdns.toks.core.common.exception.ApplicationErrorType;
 import com.tdns.toks.core.common.exception.SilentApplicationErrorException;
-import com.tdns.toks.core.domain.model.dto.UserDetailDTO;
-import com.tdns.toks.core.domain.model.entity.User;
-import com.tdns.toks.core.domain.repository.UserRepository;
+import com.tdns.toks.core.domain.user.model.dto.UserDetailDTO;
+import com.tdns.toks.core.domain.user.model.entity.User;
+import com.tdns.toks.core.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
