+++ core/src/main/java/com/tdns/toks/core/domain/quiz/model/entity/AnswerLike.java
@@ -1,4 +1,4 @@
-package com.tdns.toks.core.domain.model.entity.quiz;
+package com.tdns.toks.core.domain.quiz.model.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
