+++ core/src/main/java/com/tdns/toks/core/domain/quiz/model/entity/QuizAnswer.java
@@ -1,9 +1,7 @@
-package com.tdns.toks.core.domain.model.entity.quiz;
+package com.tdns.toks.core.domain.quiz.model.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-
-import org.springframework.data.annotation.CreatedBy;
 
 import com.tdns.toks.core.common.model.entity.BaseEntity;
 
@@ -15,13 +13,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class UserAnswer extends BaseEntity {
+public class QuizAnswer extends BaseEntity {
 
-	// TODO: 답변 포맷 논의
+	// TODO: 정답 포맷 논의
 	@Column(nullable = false, columnDefinition = "JSON COMMENT '정답'")
 	private String answer;
-
-	@CreatedBy
-	@Column(nullable = false, columnDefinition = "BIGINT COMMENT '생성자'")
-	private Long createdBy;
 }
