+++ core/src/main/java/com/tdns/toks/core/domain/user/model/entity/User.java
@@ -1,9 +1,9 @@
-package com.tdns.toks.core.domain.model.entity;
+package com.tdns.toks.core.domain.user.model.entity;
 
 import com.tdns.toks.core.common.model.entity.BaseEntity;
-import com.tdns.toks.core.domain.type.UserProvider;
-import com.tdns.toks.core.domain.type.UserRole;
-import com.tdns.toks.core.domain.type.UserStatus;
+import com.tdns.toks.core.domain.user.type.UserProvider;
+import com.tdns.toks.core.domain.user.type.UserRole;
+import com.tdns.toks.core.domain.user.type.UserStatus;
 import lombok.*;
 import org.hibernate.annotations.DynamicUpdate;
