+++ core/src/main/java/com/tdns/toks/core/domain/user/type/UserProvider.java
@@ -1,4 +1,4 @@
-package com.tdns.toks.core.domain.type;
+package com.tdns.toks.core.domain.user.type;
 
 public enum UserProvider {
     ITSELF, GOOGLE, UNKNOWN, NAVER, KAKAO, APPLE;
