+++ core/src/main/java/com/tdns/toks/core/domain/user/type/UserRole.java
@@ -1,4 +1,4 @@
-package com.tdns.toks.core.domain.type;
+package com.tdns.toks.core.domain.user.type;
 
 import org.springframework.security.core.GrantedAuthority;
